+++ src/pnet/core/process.c
@@ -11,7 +11,7 @@
 #include <pnet/core.h>
 #include <pnet/graph.h>
 #include <pnet/pnet_pack.h>
-#include <pnet/node.h>
+#include <pnet/core/node.h>
 
 static int step_process(struct pnet_pack *pack, net_hnd hnd, net_node_t next_node) {
 	net_node_t node;
