+++ src/pnet/core/repo.c
@@ -12,7 +12,7 @@
 
 #include <pnet/core.h>
 #include <pnet/repo.h>
-#include <pnet/node.h>
+#include <pnet/core/node.h>
 
 ARRAY_SPREAD_DEF(const struct pnet_module, __pnet_mod_repo);
